Fix Bug: from_dynamic not setting, and to_dynamic not honoring, __isset
Summary:[Thrift] Fix Bug: `from_dynamic` not setting, and `to_dynamic` not honoring, `__isset`.

`from_dynamic` is not setting the `__isset` marker when parsing a `dynamic`, even when the `dynamic` has that field.

As well, `to_dynamic` is not considering the `__isset` marker when choosing whether or not to render a field.

Reviewed By: juchem

Differential Revision: D3096290

fb-gh-sync-id: f2f2e06dbbc7c50c51a93731da0acdcfea8a1a8a
fbshipit-source-id: f2f2e06dbbc7c50c51a93731da0acdcfea8a1a8a