Add THeaderBody Abstraction for Thrift
Summary:After a request comes in, it is stored in an IOBuf and parsed into a
THeader and IOBuf payload pair. The payload could just be extracted from the
incoming IOBuf, but it is possible the request used compression and we applied
decompression.

In general this is fine, but there are some situations where we don't want to
apply (de)compression immediately, to defer it to another thread or use the
(de)compressed blob directly. This is possible, but introduces a danger of
using a (de)compressed payload when a user is expecting the opposite, and we
have been bitten by this type of bug before.

THeaderBody is an abstraction that wraps a compressed/uncompressed payload and
the transforms that are applied to it. It enforces very explicit conventions
for callers wanting to work with the buffer, and whether they are retrieving a
compressed or decompressed IOBuf. If the IOBuf was provided in the opposite
form, it applies (de)compression lazily and supports switching between the
two back and forth, although triggering a log to help identify wasted cpu
operations.

This changes the interface of a few Thrift APIs, particularly THeader and
server processors, so we will need to evaluate all non-library usages to make
sure nothing breaks.

Reviewed By: andrewcox

Differential Revision: D3111363

fb-gh-sync-id: dee1159b59c56be7e90f1fdd1829e5a26d48c4cd
fbshipit-source-id: dee1159b59c56be7e90f1fdd1829e5a26d48c4cd