HeaderClientChannel should respect transforms
Summary:D3111363 introduced an abstraction to combine transforms and payload
data for thrift, in order to prevent bugs and enable certain optimizations.

A bug snuck in that a THeaderBody object was constructed in
RequestChannel::clientSendT(), which is immediately after the THeaderObject was
created, but before it was populated with the write transforms. These
transforms were populated in the THeader object later down the stack in
HeaderClientChannel::setRequestHeaderOptions(), but this was not propagated to
the THeaderBody object. The net result was that the serialized data was well
formed, but the data did not use any transforms and was possibly much larger
than expected.

This diff adds a setTransforms method to the THeaderBody object that attempts
to be smart and minimizes expensive compression/decompression. This is used by
HeaderClientChannel to set the transforms before the data is serialized, fixing
the issue.

Reviewed By: andrewcox

Differential Revision: D3169206

fb-gh-sync-id: de40961bed6d028751ae4d9d0162368e080469b0
fbshipit-source-id: de40961bed6d028751ae4d9d0162368e080469b0