Let the compiler fixture test find the manifest and fixtures in its pkg-resources
Summary:[Thrift] Let the compiler fixture test find the manifest and fixtures in its pkg-resources.

This ends up being a bit more reliable because we no longer have to assume anything about the the directory layout of the fixtures and manifest, especially when they are done by a build system that manages these things in its own special way.

Note: to run these tests, you will need build rules to generate the manifest and then copy the fixture files and manifest into the test's pkg-resources.

Reviewed By: ckwalsh

Differential Revision: D3177392

fb-gh-sync-id: 1f885d55ac56654f9db1f4acdccbac7414805e3c
fbshipit-source-id: 1f885d55ac56654f9db1f4acdccbac7414805e3c