Revert THeaderBody changes
Summary: We have been seeing crashes in prod where client responses are being double decompressed, but only when the response is an exception. The THeaderBody object believes the payload is transformed, but it is not, and zlib fails. I am not certain what exactly is causing the issue, but believe it is safer to revert the changes until we fully understand the root cause.

Differential Revision: D3193160

fb-gh-sync-id: 5f2b21c1019b6483cd5dae69a6dfe49b1c93e2d2
fbshipit-source-id: 5f2b21c1019b6483cd5dae69a6dfe49b1c93e2d2