Add a new exception for queue timeouts
Summary: Since we have queue timeouts in the server actively now (hopefully), we should probably distinguish them from other error types. This adds an explicit error code. Thrift does not actually use it yet, as we want at least a quorum of clients to understand it before we do that.

Reviewed By: hiteshk

Differential Revision: D3211378

fb-gh-sync-id: 76b13a6237c524a4f95041d91ad3f10a9d55013d
fbshipit-source-id: 76b13a6237c524a4f95041d91ad3f10a9d55013d