RequestContext scope guard in Cpp2Connection
Summary: This should be the proper way to handle RequestContext in Cpp2Connection. The goal here is to set RequestContext on the thread *only* for the time interval when the thread is processing the request. Lifetime of Cpp2Request can be much longer, so it doesn't make much sense to set it in Cpp2Request constructor. There is only one place where Cpp2Request is created and it's just preceding `processor_->process()`, after which we want to reset RequestContext to the original value.

Reviewed By: haijunz

Differential Revision: D3241120

fb-gh-sync-id: 365715ab9676c4d1737ec947723fab3af53ab8e1
fbshipit-source-id: 365715ab9676c4d1737ec947723fab3af53ab8e1