Put operator< definitions in common sources in thrift tests
Summary:
[Thrift] Put `operator<` definitions in common sources in thrift tests.

The thrift compiler generates declarations for these functions in `_types.h` but does not generate definitions. One of the tests provides definitions, but not the other tests. Whether this all works or not depends on the details of the build strategy. Instead, move the definitions to a common source and link to them from all of the tests.

Reviewed By: chadparry, igorsugak

Differential Revision: D3257553

fb-gh-sync-id: ee0c7db0995962db2832003f4bb13508cb2fb755
fbshipit-source-id: ee0c7db0995962db2832003f4bb13508cb2fb755