thrift: fix an ASAN-exposed destructor-order-fiasco (DOF) bug
Summary:
In a thrift-using application, I encountered a heap-use-after-free bug as follows:

  The bad access happened on thread T2.
  The offending buffer was freed at exit-time via T0
  The "previously allocated by thread T0" stack trace pointed to THeader.cpp:63's definition:
    const string THeader::CLIENT_TIMEOUT_HEADER = "client_timeout";
  And, sure enough, it is a static-initialized non-POD whose memory is used after it is destroyed (at exit).

One fix would be to change its type to "const constexpr folly::StringPiece" and to adjust uses, appending ".str()" to each.
However, that would make each use incur the cost of a heap allocation, so this change takes the alternate approach of creating a const ref to a dereferenced and pseudo-leaked buffer.
Thus, the type of the symbol remains the same and the change is much smaller: no use requires adjustment.
I note that I was sorely tempted to change the types of the adjacent member variables (those have even fewer uses), but restrained myself and stuck with the minimal bug fix (one-thesis).
I'll fix the others as a clean-up/preventative patch separately.

Differential Revision: D3262265

fb-gh-sync-id: 05f581fe4cceeb0caee168d7c8e3f87f199c2589
fbshipit-source-id: 05f581fe4cceeb0caee168d7c8e3f87f199c2589