Fix OnewayFutureClientTest
Summary:
runFuncInQueue isn't exception safe. It is only called from async_eb
and async_eb_oneway and the lambda passed in captures the handler callback. If
an exception is thrown, the lambda will be destroyed before entering catch
block, but the catch block is still using things from the handler callback.

Fix is extend the lambda's scope. If exception is thrown, the lambda isn't
added to thread manager's queue and will be destroyed when runFuncInQueue
returns. If not, thread manager will own it.

Reviewed By: alandau

Differential Revision: D3288463

fbshipit-source-id: 9d041a5cca1d309b2c0963011c74b6f6a65221e2