Let the recv_wrapped_ generated functions wrap all exceptions
Summary:
[Thrift] Let the `recv_wrapped_` generated functions wrap all exceptions.

If a binary response comes with 0-3 bytes, then the `BinaryProtocolReader::readMessageBegin` will throw a `std::out_of_range("underflow")`. And it won't be caught and wrapped by the `recv_wrapped_` function. The likely case is that the exception bubbles up and crashes the program. No fun at all.

The reason is that the `recv_wrapped_` generated functions only catch `apache::thrift::TException`. But `folly::io::Cursor` may throw exceptions of types that don't inherit from it, most commonly `std::out_of_range`.

The easiest solution is for the `recv_wrapped_` generated functions to catch and wrap exceptions of all types.

Reviewed By: haijunz

Differential Revision: D3291825

fbshipit-source-id: 91255abb0e322e027413d4bb898cf4953867f185