in help text, group functions by service
Summary:
When the pyremote is printing the list of available methods in the help text,
group them by the service they're defined in. This is useful because it means
that the help text for the many services which extend fb303/FacebookService
will be easier to read.

This change is both to the compiler and the python runtime.

Reviewed By: alandau

Differential Revision: D3252199

fbshipit-source-id: 59e9c95f00bb78d2a554f5ee822245e3c4155761