Use HeaderClientChannel::newChannel everywhere
Summary:
[Thrift] Use `HeaderClientChannel::newChannel` everywhere.

Instead of `new HeaderClientChannel`. The factory static method does that and wraps it in a `unique_ptr` with the proper destructor type. So that caller code does not need to do so.

In some places, do some broader rewrites if better techniques are available. In particular, in a variety of test cases, `ScopedServerInterfaceThread::newClient` is the best technique.

Reviewed By: igorsugak

Differential Revision: D3292339

fbshipit-source-id: 039b37c29b39d10b90820fc1a8bc935b726fcdcc