Fix bit counting for narrow types
Summary:
Fixing the UB when left-shifting in D3310688 introduced a bug when calling
bitsNeeded() for signed types narrower than an int. The previous (UB) code
did cause sign extension to kick in, so the result of (x ^ (x << 1)) did
have all MSBs cleared. The new code using unsigned types will obviously
not do that, so we have to explicitly narrow the argument to findLastSet()
to count only bits that are within the width of the original type.

Differential Revision: D3316303

fbshipit-source-id: d8d54d907cf29d3dd774ff4cd540d26ad3fc5ca1