thrift: expose the TServerObserver interface to Python
Summary:
The TServerObserver interface allows an observer to be notified of a number
of interesting events in the lifecycle of Thrift client requests.  This can
be used to generate a number of useful stats, including information about
overall load and throughput.

This adds a facility to expose the TServerObserver to Python, by providing
an interface class that adapts from the C++ side to method calls on the
Python object supplied.

This adapter is only added when a Python observer is added to the Thrift
server, so there is no performance cost if the new facility is not used.

Reviewed By: yfeldblum

Differential Revision: D3330027

fbshipit-source-id: 6e1f1a2ad7d5e1766243512f0628b00783d1652d