Reduce THeader zlib untransform memory allocation
Summary:
We allocated a fixed 1K buffer in a loop when decompress zlib, this
could be lots of allocations if the decompressed message is big. This uses the
size of the compressed chunk so it needs fewer allocations.

Reason to use bigger min allocation size is this can further reduce the number
of allocations needed, but could waste memory, so only increase it for large
messages.

Reviewed By: alandau

Differential Revision: D3354759

fbshipit-source-id: 3306f634289256bad44b22d83d5469fe5e5d23fa