Implement enumFromTo for Haskell enumerations
Summary:
[Thrift] Implement `enumFromTo` for Haskell enumerations.

And the rest of `Enum`, namely, `succ`, `pred`, `enumFrom`.

We do not implement `enumFromThen` or `enumFromThenTo` since, if the enum values do not correspond to the sequence of integers, there is no correct implementation.

Motivation: If one want a list of all valid enum values, one might want to use `[minBound..maxBound]`. This diff makes that work correctly when the enum values do not correspond to the sequence of integers.

Examples may be seen in:

    thrift/compiler/test/fixtures/qualified/gen-hs/Module0_Types.hs
    thrift/compiler/test/fixtures/qualified/gen-hs/Module1_Types.hs

Reviewed By: zilberstein

Differential Revision: D3384552

fbshipit-source-id: e7ea6e06cb638cdd69ecb62f3d3b6c90bae08367