accept json serialization options in thrift cpp readFromJson
Summary:
thrift cpp1 uses parseJson() in readFromJson, without any options. the thrift struct can contains some large double values that got converted to a large int string and cause troubles.

the root cause is, thrift cpp1 readFromJson() first deserialize the string, get a dynamic, and the serialize part of it so that the member fields can parse it again. in other words, it's string -> dynamic -> string -> dynamic. during the second conversion, some large doubles will become int strings as configured by kConvMaxDecimalInShortestHigh. then in the third conversion, this big int is either causing precision problems or overflow problems.

double_fallback in folly::json::serialization_opts can solve this problem. and we need a way to pass it in. this diff does just that.

Reviewed By: alandau

Differential Revision: D3391761

fbshipit-source-id: 147ec045075972fd4c10e4b6920fbedf9e000fcb