buckification: add minimal node/js support
Summary:
This adds minimal support for node/js rules via an implementation
purely in the buckification macros.  This should support existing
use cases in fbcode so that node/js projects have a way to move now
that fbbuild is deprecated, but there aren't plans to fully support/
maintain node/js support on the Build Infra side.

Reviewed By: markisaa

Differential Revision: D3406705

fbshipit-source-id: 51489a0cedc39af240fa32641c79bb8fc7b7ec17