For a frozen range of trivials, provide direct access
Summary:
[Thrift] For a frozen range of trivials, provide direct access.

In Frozen 2.

After freezing a `std::vector<float>`, provide direct access as a `folly::Range<const float*>`.

This may seem like a dirty hack, but the Frozen 1 library provided this. Code that uses this hack with Frozen 1 may be made easier to port to Frozen 2 if this dirty hack were provided in Frozen 2 as well.

Reviewed By: ddrcoder

Differential Revision: D3423179

fbshipit-source-id: 09a9437cff45ba89eb1dfc5e8a18601299872a28