Drop boost::totally_ordered as a base class for cpp2 structs
Summary:
[Thrift] Drop `boost::totally_ordered` as a base class for cpp2 structs.

Having this as a base class has a pro and a con.
Pro - It makes it a bit easier to have comparison operators beyond `==` and `<`. Upside is very limited in practice.
Con - it introduces functions in boost into the ADL overload resolution sets, easily leading to ambiguity when `namespace std` or `std::move` is in scope. Ambiguity is a compiler error. Downside is common when upgrading existing code to use cpp2.

Example of the con in practice:

```
using namespace std;
// 1000s of lines ...
void foo(MyThriftStruct v0) { // a cpp2-generated type from a thrift struct
  auto v1 = move(v0); // error: ambiguous between std::move and boost::move
}
```

Reviewed By: alandau

Differential Revision: D3454767

fbshipit-source-id: 83059fef86e875787249dcaa99cc17f504a55168