Fix the error of folly_dynamic when type_class is binary
Summary:
This diff fixed the error of folly_dynamic when type_class is binary.

Basically `binary` is added to `type_class` but `folly_dynamic` does not handle it. My project in `neteng/fboss/tools/faat` ( D3494819 ) breaks because of this change, because `fboss/agent/if/ctrl.thrift` have some binary fields.

Reviewed By: juchem

Differential Revision: D3518454

fbshipit-source-id: 9316d0c7bf2af85510cb501964c402dee283233b