thrift: need to ensure that we add -lnuma in the opensource build
Summary:
This is fixing a problem that I introduced recently; the default
action for `AC_CHECK_LIB` when it locates the library is to add it to the
library flags.  But since we're now running some headers checks in that case,
the library flags are not added by default and linux builds fail to link :-(

This diff seeks to fix this problem by explicitly adding them back in in the Makefile.am when we know that libnuma is present.

Reviewed By: alandau

Differential Revision: D3585256

fbshipit-source-id: e5039446f92a1191f5e60f23e368ade51cf2c1ca