Fix scheduling bug in TimerManager
Summary: The current conditional will always evaluate to false, because if a task is inserted with lower timeout, it will be returned by begin() and the comparison will degenerate to timeout < timeout, which is always false.  Since this is performed while a lock is held the insertion can simply be moved to after the comparison.

Reviewed By: yfeldblum

Differential Revision: D3565168

fbshipit-source-id: c94d46237363f03b8cd9f6cbce13a70da705f950