Fix handling of EOF events
Summary:
On Thrift server side, a callback passed to
getDuplexClient<Client>.getChannel().setCloseCallback() was not
executed when the Thrift client (the remote side) dropped the
connection. This diff fixes that.

This diff also revealed problems about object lifetime that ASAN
reported as heap-use-after-free. Since DuplexChannel stores its
DuplexClientChannel, DuplexServerChannel and DuplexCpp2Channel
objects as shared_ptr, which are accessible to the user via
getter methods, there is no guarantee that the DuplexChannel
object outlives those three child objects. This diff makes sure
the child objects don't access the DuplexChannel object if it
has been destroyed in the meantime.

Reviewed By: alandau

Differential Revision: D3555137

fbshipit-source-id: b1c3614cfa014684fa779dac4b872f3874c1895e