Allow multiple peekers to be used in wangle::Acceptor.
Summary:
Previously only one PeekCallback could be used at once, and it was chosen by the last derived class to override startHandshakeManager(). This made it difficult to configure different peekers at different levels of the code, or to configure the same peeker in different derived classes of wangle::Acceptor.

With this change multiple peekers are added as the acceptor is initialized, and are visited in order when a new connection is received. Instead of 1 peeker having to make a final decision on which HandshakeHelper to start, it can defer to the next peeker. The final peeker defaults to SSL.

Reviewed By: siyengar

Differential Revision: D3640418

fbshipit-source-id: 1cbbe0b582899a63b8ac8d00a1272c4f0b2f6c34