Let debug_equals over unordered maps and sets be insertion-order insensitive
Summary:
[Thrift] Let `debug_equals` over unordered maps and sets be insertion-order insensitive.

Currently, over unordered maps and sets, it is sensitive to insertion order. Generally, only maps and sets with entries inserted in the same order will compare equal.

This is undesired. It is better to compare maps and sets for equality of contents, without requiring equality of insertion order as well.

Reviewed By: juchem

Differential Revision: D3689082

fbshipit-source-id: c1c00da8e96dd1370b3a830dcc4f38a153dd3912