Fix union deserialization for cpp and cpp2
Summary:
[Thrift] Fix union deserialization for cpp and cpp2.

When `readFieldBegin` yields a field type of `T_STOP` (or is assumed to), then `readFieldEnd` should not be called. This is how struct deserialization works.

The bug here is that `readFieldEnd` was being called even though `readFieldBegin` yielded a field type of `T_STOP` after reading a union's field, and the call to `readFieldEnd` would consume an extra closing brace `}` unmatched by any preceding calls to `readFieldBegin`. Then a subsequent `readFieldBegin` would run out of buffer and would throw.

This affects the json (tested) and possibly the simple-json (suspected) protocols. It does not affect the binary or compact protocols because `readFieldEnd` is a no-op.

Reviewed By: alandau, juchem

Differential Revision: D3474801

fbshipit-source-id: 5f6fe0e42e62518107162c1ad2f350256190ff34