suggest fix for ref struct when the struct has required field
Summary:
in cpp1, when a field is annotated as ref (pointer), and this field is a struct, and also
the struct has required field. During deserializing of this field, it could throw "required
field was not found in serialized data" when the serialized field is a nullptr.

For example, we have following struct:
```
struct Test {
  1: required string str;
}

struct HasProblem {
  1: optional Test test (cpp.ref = "true");
}
```
Then we have HasProblem.test as nullptr, and serialize, in deserialization side, will throw
above exception.

This diff is trying to fix the cpp1 as the exactly same behaviour as cpp2.(yes, cpp2
doesn't have this problem.)

Reviewed By: yfeldblum

Differential Revision: D3715079

fbshipit-source-id: 6f887851bb9b5c0e2f8ab2fe335571d42cec8653