Fewer static arrays and template instantiations in legacy reflection
Summary:
[Thrift] Fewer static arrays and template instantiations in legacy reflection.

Fewer static arrays, theoretically, because there were cases, such as when a structure type was the element type of a list, where we might have a static array for both the structure name and the null-terminated structure name.

Fewer template instantations, theoretically, because null-terminating at compile time has to add a template instantiation.

Reviewed By: dymk

Differential Revision: D3715031

fbshipit-source-id: 61174e83f76d8249ed22972f4a46ca82d8307d20