specialized read for bools, to accont for std::vector<bool> proxy
Summary:
`list<bool>` doesn't work with the templated serializer, because
the serializer assumes that all containers used in Thrift actually
act like containers. `std::vector<bool>` is, of course, not a container,
This adds an overload on the primitive writing `protocol_methods`
for `bool`, to accept the proxy bool that `std::vector<bool>` provides.

Reviewed By: yfeldblum

Differential Revision: D3738218

fbshipit-source-id: ee376d0e42fecf65fcb76ce97b5a135f08369b94