Disable apache::thrift::merge on cpp2-generated types
Summary:
[Thrift] Disable `apache::thrift::merge` on `cpp2`-generated types.

The problem is that the default implementation is wrong for `cpp2`-generated struct types, but `cpp2` does not generate ADL overloads. So when upgrading code that ADL lookup of `merge` from legacy `cpp` to `cpp2`, the default implementation would be found; the behavior of the code would implicitly be changed and there would be no warning.

An easy solution is to disable `merge` explicitly for all types known to be `cpp2`-generated. There is no public API for this, but we can use a known base class that is an implementation detail of `cpp2` to detect `cpp2`-generated types.

Also add some comments explaining the situation, with a pointer to `apache::thrift::merge_into`.

Reviewed By: philippv

Differential Revision: D3740980

fbshipit-source-id: bde0e81dd2e52e42a6d3a5e5553d701c823d805c