Fix performance issue while serializing a struct contains a lot double as Json
Summary:
folly::to<std::string>(double), will first reserve 24 bytes (25 bytes if it's a negtive value), and do toAppend.
our std::string will do a malloc if the reserve bytes is larger than 23 bytes, hence each time serialization a double, it
will have a malloc and free happen, this is quite expensive especially in a high load service.

By fixing this as this diff suggests, in a single thread testing 10000 times serailizing a struct contains 100 double,
it saves 41% wall time. (on an opt build).

previous : 0.187114 second
after : 0.13255 second

(Notice, this is on one single thread, hence the malloc and free has the benefit from thread local arena. in real system,
we should be able to save more).

Reviewed By: yfeldblum

Differential Revision: D3725723

fbshipit-source-id: 2aaf8736d852f01dc75b226a3764aab55649b4e1