Avoid std::auto_ptr in fixtures
Summary:
[Thrift] Avoid `std::auto_ptr` in fixtures.

There is no support anywhere for that smart-pointer type, so don't use it.

Thrift allows specifying which smart-pointer type is to be used with a given `ref` field, but only `std::unique_ptr` and `std::shared_ptr` are known smart-pointer types and handling is implemented for those smart-pointer types only.

Let the fixtures reflect the reality of what Thrift actually supports, and avoid any smart-pointer types in the fixtures that are not actually implemented in Thrift.

Reviewed By: markisaa

Differential Revision: D3747587

fbshipit-source-id: 9d8aec60eb239209e650f1d7686fa75c58d71c91