Provide default container-traits for std-like containers
Summary:
[Thrift] Provide default container-traits for std-like containers.

It seems something of a senseless pain to specialize the container traits with so many methods when the container type being specialized over is designed to have the same interface as a container in the standard.

Provide a sensible default suite of container traits types which work over container types that mimick standard containers, so that specializing them for non-standard but standard-mimicking containers is painless.

And use them for the container traits specializations for the existing standard and Folly types.

Reviewed By: markisaa, juchem

Differential Revision: D3757066

fbshipit-source-id: bb1a2848ec1a44042c3e3d8db3460e494b10607b