Move the definition of merge over unique_ptr down
Summary:
[Thrift] Move the definition of `merge` over `unique_ptr` down.

This solves a problem where one of the fixtures would not actually build because the definition of `merge` over `unique_ptr` would call, via ADL, an overload of `merge` that was declared afterwards. In particular, in this fixture, the overloads for `vector` and `set`. The fixture in question is `thrift/compiler/test/fixtures/refs/src/module.thrift`.

It is required for ADL that all candidates lexically be declared before they are called, they may be forward-declared before all call-sites and only defined later.

Only the overload over `unique_ptr` is moved because that is the only one that is required to be moved in this particular case.

Reviewed By: markisaa

Differential Revision: D3748149

fbshipit-source-id: cb710b758e6420e5835e7a6cd161398cea67e8e0