level order insertion of elements into map, given the right circumstances
Summary:
For large maps, the cost of red-black tree rotation is fairly costly. This
diff implements a change to insert the contents of a map in level-order to
avoid RB tree rotations for sufficiently large maps.

Probably the hackiest thing I've done in the past 5 weeks. Lets see if it's
worth it.

Reviewed By: juchem

Differential Revision: D3604369

fbshipit-source-id: 21695506cbe262f374080c552dd114e34b00cd04