Prevent attempted undefined static_cast of t_typedef& to t_struct&
Summary:
When the thrift parser encounteres a type that has not yet been defined, the type is stored as t_typedef instead of
as its not yet known actual type.  After parsing is complete, the code generator looks up the type definition as needed.  In the existing code
path modified by this diff, an attempt is made to static_cast t_typedef& to sibling type t_struct&, yielding undefined behavior.  This has gone
undetected because the current observed behavior is to do nothing; the type continues to be handled as  t_typedef&.  This reference is then immediately converted
to t_struct& by _get_true_type.  This call of _get_true_type merely repeats the conversion performed 11 lines earlier and may be
eliminated along with the undefined static_cast.

Reviewed By: yfeldblum

Differential Revision: D3756450

fbshipit-source-id: dd48a3b53997be0f5bcfc7fe0570f2a361a45397