Treat all Dynamic strings as binary
Summary: Yesterday's try to introduce a separate `bin` field in Dynamic wass ill-conceived because there's no equivalent folly::dynamic type. This causes Python clients to be unable to read certain dynamic fields. Instead, we're switching everything to `binary` which means the same thing in C++ but will be raw bytes on Python 3.

Reviewed By: haijunz

Differential Revision: D3774894

fbshipit-source-id: d073a8e7e9805f20584ced14af666501df251ebb