Fix CloneableIOBufTest
Summary:
The intention of this test is it binds to operator='s rvalue overload,
but it actually binds to the const reference overload. In this context the
parameter of rvalue overload is a rvalue reference, not a universal reference.
So it is an exact match but const lvalue reference needs an implicit
conversion. I am not sure why compiler makes this choice here.

The constructor overloads work as expected. Did I miss anything obvious?

Reviewed By: yfeldblum

Differential Revision: D3783438

fbshipit-source-id: 5fecb8c8d59146598a82198d1f7377d1b4cdcfab