Add convention to dump strings
Summary:
Create a convention that dumping a string creates a new object with a "value"
key and the string as the value. Why would one want to do this when mustache
makes it easy to include a string or an array of strings in a map?

The dump_elems function handily adds first? and last? attributes to each element
it dumps in a list. It calls the dump() method for each item in the list. By
overloading dump to accept a string, we get to take advantage of the magic
first? and last? values with consistency and future-friendliness.

Reviewed By: yfeldblum

Differential Revision: D3828488

fbshipit-source-id: 26ee703a0319212ef8dccc94172cbc94b74a2db2