Add generated service wrapper code
Summary:
We need (fairly lightweight) wrappers of the C++ interfaces that are aware of
the PyObject handler that contains the Python coroutines that contain
user-written code. These templates set up the boilerplate for the .h and .cpp
files for those wrappers.

I haven't included the function definitions here; I will do that in an upcoming
diff.

Reviewed By: chadparry

Differential Revision: D3830477

fbshipit-source-id: d0cb5e59fd6984c655e725ab5e3e1911a4e1dbdf