Avoid a circular reference
Summary:
We need to use a weak reference to hold the Thrift server while it is starting up, to avoid a circular reference. Once the `ScopedServerThread::Helper` is unregistered as an event handler from the Thrift server, then we can (and should) go back to a owning smart pointer.

See https://our.intern.facebook.com/intern/sevmanager/view/s/140089/.

Differential Revision: D3823512

fbshipit-source-id: 789111ea92b2ac727aa2dc95d0bf8563efeda7b3