Refactor render and write sequence to its own method
Summary:
As I started writing new rendering functions, I realized that this pattern was
going to be really common. Indeed, it may be worth it to put it on the
superclass. This diff pulls the duplicate code into its own method. Later diffs
will take advantage of the new method.

Reviewed By: chadparry

Differential Revision: D3845024

fbshipit-source-id: 4d193804af7076cbc7f6ef0f916e766bcee05c55