Generate Cython wrappers of generated C++ structs
Summary:
The end goal is to store data in C structures and convert attributes to Python
as late as possible. This diff creates the Python wrappers of the C structures.
When a primitive attribute is accessed, it is returned directly, allowing Cython
to do the correct conversion. When a struct attribute is accessed, a Python
wrapper of the underlying C struct is created, but no data is converted to a
PyObject yet. This wrapper is memoized on the class so that the wrapping step
only has to happen once.

This diff relies on all references to complex objects being shared_ptr in the
C++ generated code. Haijun is working on this in t12998795.

Reviewed By: chadparry

Differential Revision: D3848239

fbshipit-source-id: 457e48f491e11382f9f96c39de65ecd8e1b3f4a4