Fixing heap-use-after-free bug in ContextStack
Summary:
ContextStack had a heap-free-after-use bug that may have been causing a memory leak.
The char* in ContextStack was getting deleted outside of this class, but was still being used as
a key into a map. So this was likely creating map entries with freed memory as keys.

This fix just changes it to a std::string which should make a copy of the char* and get freed when the
ContextStack is freed

Reviewed By: meyering

Differential Revision: D3853356

fbshipit-source-id: 1641bde705bb77ba6031bca0d16056980a16cdd8