Move fbthrift to a proper maven build (while still maintaining BUCK build in parallel)
Summary:
This diff set ups "proper" maven build for fbthrift and also attempts to maintain parallel buck build. As of now it's FYI diff (I still need to fix all fbcode buck builds that depend on fbthrift).

Initial goal here is to make fbthrift "maintainable" by unifying build and release strategy.
**Impact here** is to have: healthy, maintainable and java-idiomatic setup for fbthrift development and release proccess.
Final state (state we're trying to achieve) is:
1. fbthrift is built only with maven
2. fbthrift is released to Nexus via maven-release-plugin in 1 easy step
3. all buck projects depend on fbthrift via fbcode/java-third-party
4. all tests in fbthrift (including original apache thrift tests) work properly
5. continuous integration is set up for fbthrift
6. one can work with fbthrift in IDE

Immediate steps after this diff:
1. Update all buck projects that depend on fbthrift to depend on new targets file
2. Release fbthrift to Nexus
3. Update all buck projects that depend on fbthrift to use java-third-party

Now, the structure of new fbthrift maven build is:
**thrift-root** is parent module that defines reactor build
**thrift** is module that contains all fbthrift code (no edits)
**thrift-tests** is module that contains original apache thrift tests, that module is currently disabled because those tests were not maintained
**thrift-directserver** is a module for direct server.

In comparison to original project structure there are 2 important changes:
1. directserver now lives in a separate module outside of thrift source tree
2. generated code is now checked into source tree (for now that's the best option)

BUCK build (which is about to be deprecated) does not use checked-in generated source code, while maven build **does**. This approach is dictated by javafoundations previous experience of migrations from BUCK to maven.

This is FYI diff, it is not intended to be committed in it's current form, but the //structure// of the build is final.

Reviewed By: ryandm

Differential Revision: D3864458

fbshipit-source-id: 80c972a68292eec9c4f321cede0252e929ed7985