Create c++ implementation to call into cython
Summary:
The whole point of this C++ wrapper is to have a small file that knows that it
owns a PyObject and to pass that PyObject into a cython handler. I haven't
actually tested that which I am trying to do, but the goal is to pass a promise
down to Cython, then have Cython call into a coroutine and set the promise value
whenever the coroutine finishes.

Reviewed By: cjtenny

Differential Revision: D3852150

fbshipit-source-id: 07e7007c5cc275ce17664af6b9dc3dfb5eff508c