Wrap the C++ Wrapper with Cython
Summary:
Yes, wrapping the wrapper. The c++ wrapper is responsible for calling back into
cython functions. As such, we don't need to expose most of the handler's
interface, just the factory to construct it. (BTW, that factory is necessary
because, as far as I can tell, Cython is not able to polymorphically determine
that a wrapped subclass is an instance of a wrapped superclass).

Reviewed By: cjtenny

Differential Revision: D3865546

fbshipit-source-id: 8b8612cad903cdc3fa0e53ee8177009889a4c99b