Make template and file naming more consistent
Summary:
As I was writing more templates, I realized I was expecting names to be
different. So I figured I'd change them before it's too late.

I removed the cy_ prefix because the fact that a file is a pxd or pxi is enough
to identify it as a cython file.

I made all files that can contain services have the plural "services" instead of
the singular "service".

Reviewed By: cjtenny, chadparry

Differential Revision: D3868060

fbshipit-source-id: cf0a937c9907ce181e9d9783c13d8948e52e5078