Create promise wrappers for each function return type
Summary:
The Cython callback code needs to be able to return a C++ Promise, but that
promise needs to be passed into Python code inside a coroutine so that setValue
can be called correctly. Unfortunately, there doesn't seem to be a way to attach
a generic type to a cdef class, so we need Python wrappers for each type we
might return. (Note: There is experimental work towards generics in Cython's
Fused Types implementation. However, it does not yet support fused attributes on
extension types yet).

Reviewed By: cjtenny

Differential Revision: D3883988

fbshipit-source-id: 58415a79adc33b0973f92f243f0f9c20c59260ca